% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{import_data}
\alias{import_data}
\title{Import Data}
\usage{
import_data(dataset)
}
\arguments{
\item{dataset}{A character vector from a controlled list of datasets.}
}
\value{
The identified dataset will be loaded into the global R
        environment.
}
\description{
Imports Datasets.
}
\section{What it does}{

As the name suggests, this function downloads data from the package.
For most datasets, you can simply use the base `data()` function,
but some datasets are stored across multiple files, and this function
automatically compiles them.


Valid dataset names include:
"behn_A27301",
"behn_peer_corpus",
"conceptual_change",
"doc_metadata",
"eebo_network",
"footprints",
"geo",
"keyword_doc",
"kwic",
"locke_A48901",
"locke_peer_corpus",
"place_doc",
"place_year",
"term_doc"
}

\examples{
import_data("kwic")


}
